# docs/swagger.yaml
definitions:
  main.FileUploadResult:
    properties:
      id:
        type: string
      presigned_url:
        type: string
      public_url:
        type: string
    type: object
  main.GenRequest:
    properties:
      aspect_ratio:
        description: option
        type: string
      expand_prompt:
        description: option
        type: boolean
      image_end_url:
        description: option, uploaded refer image url
        type: string
      image_url:
        description: option, uploaded refer image url
        type: string
      user_prompt:
        description: option
        type: string
    type: object
  main.UploadReq:
    properties:
      url:
        description: support public url & base64
        type: string
    type: object
  main.Video:
    properties:
      height:
        type: integer
      thumbnail: {}
      url:
        type: string
      width:
        type: integer
    type: object
  main.VideoTask:
    properties:
      created_at:
        type: string
      estimate_wait_seconds: {}
      id:
        type: string
      liked: {}
      prompt:
        type: string
      state:
        description: '"pending", "processing", "completed"'
        type: string
      video:
        $ref: '#/definitions/main.Video'
    type: object
info:
  contact: {}
paths:
  /luma/generations:
    post:
      consumes:
      - application/json
      parameters:
      - description: submit generate video
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/main.GenRequest'
      produces:
      - application/json
      responses:
        "200":
          description: generate result
          schema:
            items:
              $ref: '#/definitions/main.VideoTask'
            type: array
      summary: Submit luma generate video task
  /luma/generations/:
    get:
      consumes:
      - application/json
      parameters:
      - description: page offset
        in: query
        name: offset
        required: true
        type: string
      - description: page limit
        in: query
        name: limit
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: video tasks
          schema:
            items:
              $ref: '#/definitions/main.VideoTask'
            type: array
      summary: Get luma generate video task
  /luma/generations/:task_id/extend:
    post:
      consumes:
      - application/json
      parameters:
      - description: extend task id
        in: path
        name: task_id
        required: true
        type: string
      - description: submit generate video
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/main.GenRequest'
      produces:
      - application/json
      responses:
        "200":
          description: generate result
          schema:
            items:
              $ref: '#/definitions/main.VideoTask'
            type: array
      summary: Submit luma extend generate video task
  /luma/generations/{task_id}:
    get:
      consumes:
      - application/json
      parameters:
      - description: fetch single task by id
        in: path
        name: task_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: video single task
          schema:
            $ref: '#/definitions/main.VideoTask'
      summary: Get luma generate video task
  /luma/generations/{task_id}/download_video_url:
    post:
      consumes:
      - application/json
      parameters:
      - description: fetch by id
        in: path
        name: task_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: url
          schema:
            type: object
      summary: Get video url without watermark
  /luma/generations/file_upload:
    post:
      consumes:
      - application/json
      parameters:
      - description: Upload image params
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/main.UploadReq'
      produces:
      - application/json
      responses:
        "200":
          description: upload result
          schema:
            items:
              $ref: '#/definitions/main.FileUploadResult'
            type: array
      summary: Upload image to luma
  /luma/subscription/usage:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: subscription info
          schema:
            type: object
      summary: Get current user subscription usage
  /luma/users/me:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: user info
          schema:
            type: object
      summary: Get current user info
swagger: "2.0"